feat: improved yayy
